Retake task4           -Move functions for read and write to file to separate file -Handle issues when sending request related to appropriate issues
